BE security + bug fixes

+++ server/controller/adminControl.js
@@ -2,8 +2,17 @@ import { userModel } from "../models/Users.js";
 import { ishmtFileModel } from "../models/IshmtFile.js";
 import { paymentFileModel } from "../models/PaymentFile.js";
 import { formModel } from "../models/Form.js";
+import jwt from "jsonwebtoken";
 
 const userListControl = async (req, res) => {
+  const token = req.body.token;
+  const userID = jwt.decode(token, process.env.JWT_SECRET);
+  const user = await userModel.findById(userID.id);
+  if(!user.isAdmin)
+  {
+    console.log("Unautorized Access!")
+    return res.json({message: "Access Denied", success : "false"})
+  }
   try {
     const data = await userModel.find({ isAdmin: "false" }, { userEmail: 1 }).sort({ userEmail: 1 });
     console.log("List Fetched Successfully");
@@ -19,6 +28,15 @@ const userListControl = async (req, res) => {
 };
 
 const userDeleteControl = async (req, res) => {
+
+  const token = req.body.token;
+  const userID = jwt.decode(token, process.env.JWT_SECRET);
+  const user = await userModel.findById(userID.id);
+  if(!user.isAdmin)
+  {
+    console.log("Unautorized Access!")
+    return res.json({message: "Access Denied", success : "false"})
+  }
   try {
     await userModel.findByIdAndDelete(req.body.userID);
 
@@ -50,6 +68,14 @@ const userDeleteControl = async (req, res) => {
   }
 };
 const userDownloadControl = async (req,res) =>{
+  const token = req.body.token;
+  const userID = jwt.decode(token, process.env.JWT_SECRET);
+  const user = await userModel.findById(userID.id);
+  if(!user.isAdmin)
+  {
+    console.log("Unautorized Access!")
+    return res.json({message: "Access Denied", success : "false"})
+  }
   const usersData = await formModel.find({}).sort({ fullName: 1 });
   return res.json({usersData, success:"true"});
 };
